Rename dsl class
