+++ lib/heel.rb
@@ -5,8 +5,7 @@ module Heel
   require "heel/server"
   require "heel/registry"
   require "heel/response"
-  require "heel/bot"
-  require "heel/bot_v2"
+  require "heel/dsl"
   require "heel/bot_manager"
   require "heel/new_bot"
